Make it easy to swap out and implement different transcription models
