+++ src/main/config.ts
@@ -1,5 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 
+import { TranscriptionEngine } from './handlers/media/transcriptionEngines';
+
 // Whether to use the sample transcription instead of actually transcribing.
 // Useful for testing.
-export const USE_DUMMY = false;
+export const TRANSCRIPTION_ENGINE = TranscriptionEngine.DEEPSPEECH;
